pyproject.toml
[project]
name = "py-tool"
version = "0.1.0"
description = "GitLab提交查询工具"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.29.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "itsdangerous>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]